requirements.txt
Pillow==9.3.0
PyYAML==6.0
